User Registration & Login Functionality
Implemented a basic user registration and login system.

Usernames and passwords are stored in the database and upon a login request, the entered username is looked up in the database and the stored password is checked against the one provided. If the credentials match, the user is logged in and passed to a session page.

The registration checks the database for existing users with the same username trying to be registered. If none exists a user is created with the entered username and password.